backend/requirements.txt
Flask[async]
flask-cors
gunicorn

pytest-asyncio